arch-arm: New version of decoding for AdvSIMD
This commit is rewriting the decodeFpAdvSIMD function
using the new bit pattern matcher utility.
The purpose is to have a more tabular format for the
decode tree to more closely match the Arm pseudocode
specification.

As we are trying to match the machInst with all the instruction
group matchers until one is found, the time complexity is now
O(n). This means the flatter the instruction group
is, the more it might potentially take to decode an instruction.
Luckily the checking is as simple as an AND & COMPARE so even
with a big instruction pool like the SIMD one, the overhead is
negligible (and IMHO because of the sometimes complex decoding
logic we have been using, it might even make it better)

This has been tested with the following benchmark, which is
executing an instruction FMADD from the last entry in the decoding
table Fp data processing 3-source (hence testing the
"worst case scenario"):

int main()
{
    uint32_t prog[PROG_SIZE] = {0};

    std::fill_n(prog, PROG_SIZE, FMAD);
    prog[PROG_SIZE - 1] = RET;

    typedef double (*fmadd_func)(void);
    fmadd_func func = (fmadd_func)&prog[0];

    for (int idx = 0; idx < 10000; idx++)
        func();

    return 0;
}

The benchmark (run with atomic in SE mode) runs PROG_SIZE * 10000 = 1bn
FMADDs and shows no difference between introduced by the current
patch (runtime is equally ~14minutes in my machine)

Change-Id: I63709626baa706ad9c664abc753215843aa08351
Signed-off-by: Giacomo Travaglini <giacomo.travaglini@arm.com>