util: Add gem5_bridge driver module
The module adds a device called gem5_bridge that allows the m5 utility
to mmap to the gem5ops MMIO range for the given architecture. The deivce
has user read/write permissions, so root access is no longer required to
perform gem5ops. The device has configurable parameters for the base and
size of the MMIO range, as each architecture in gem5 expects a different
range.

One caveat is that the module needs to be built for the same exact
kernel version as it is being inserted into. The easiest way to achieve
this is by extracting the vmlinux kernel from the diskimage you have
built the module into and use that kernel with gem5 during simulation.