misc: Fix Scons tag resolution
This portion of the code is used to filter the list of
all sources to figure out if it contains a certain tag.
We should not resolve the tags of the source list being
filtered, otherwise we do not filter anything.

In other words, when parsing a file F, which requires tag
T, if another file F2 contains tag T2, which implies tag
T, then F2 is should not be a dependency of F.

As a side effect, improved the printing functionality of
a couple of types.

Signed-off-by: odanrc <odanrc@yahoo.com.br>