arch-arm: Allow HFGR/WTR_EL2 matching
At the moment the callback only checks if the corresponding bit
is set, and traps to EL2 accordingly (for example: HFGRTR_EL2.MAIR_EL1).
However the register contains some n-prefixed bits whose meaning is:
"trap if bit is NOT set (=0)"

We therefore add an extra template parameter, which is a matching
value defaulting to one. Existing "trap if bit is set" cases are
therefore untouched. New bits will have to set the third param to
zero

Change-Id: If73df310dd48e82435faadd90b8fb4e9f30dd4c8
Signed-off-by: Giacomo Travaglini <giacomo.travaglini@arm.com>